backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
sqlmodel>=0.0.22
psycopg[binary]>=3.1
python-multipart>=0.0.9
alembic>=1.13
